Can create tools from spec json
